+++ sdccc/src/main/java/com/draeger/medical/sdccc/messages/mapping/MessageContent.java
@@ -16,6 +16,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import javax.annotation.Nullable;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.ElementCollection;
@@ -75,6 +76,9 @@ public class MessageContent {
     private String uuid;
     private boolean isSOAP;
 
+    @Column(nullable = true)
+    private String sender;
+
     /**
      * This will be used by hibernate when creating the POJO from database entries.
      */
@@ -94,6 +98,7 @@ public class MessageContent {
      * @param actions              ws addressing actions
      * @param uuid                 identifier for ensuring, that a message was written to the database
      * @param isSOAP               shall be true if a SOAP envelope was found and false otherwise
+     * @param sender               the IP address of the message's sender or null if it could not be determined.
      */
     public MessageContent(
             final String body,
@@ -105,7 +110,8 @@ public class MessageContent {
             final List<MdibVersionGroupEntity.MdibVersionGroup> mdibVersionGroups,
             final Set<String> actions,
             final String uuid,
-            final boolean isSOAP) {
+            final boolean isSOAP,
+            @Nullable final String sender) {
 
         this.body = body;
         this.direction = direction;
@@ -115,6 +121,7 @@ public class MessageContent {
         this.actions = actions;
         this.uuid = uuid;
         this.isSOAP = isSOAP;
+        this.sender = sender;
 
         this.messageHash = MessageUtil.hashMessage(this.body);
         this.scheme = communicationContext.getTransportInfo().getScheme();
@@ -213,5 +220,9 @@ public class MessageContent {
 
     public List<MdibVersionGroupEntity> getMdibVersionGroups() {
         return this.mdibVersionGroups;
+    }
+
+    public String getSender() {
+        return this.sender;
     }
 }